Refactor: rename annif.hit module to annif.suggestion . Part of #267
