+++ annif/backend/dummy.py
@@ -1,7 +1,7 @@
 """Dummy backend for testing basic interaction of projects and backends"""
 
 
-from annif.hit import SubjectSuggestion, ListSuggestionResult
+from annif.suggestion import SubjectSuggestion, ListSuggestionResult
 from . import backend
