+++ tests/test_eval.py
@@ -3,7 +3,7 @@
 import numpy as np
 import annif.corpus
 import annif.eval
-import annif.hit
+import annif.suggestion
 
 
 def test_precision_at_k():
@@ -92,14 +92,14 @@ def test_evaluation_batch(subject_index):
 
     gold_set = annif.corpus.SubjectSet.from_string(
         '<http://www.yso.fi/onto/yso/p10849>\tarkeologit')
-    hits1 = annif.hit.ListSuggestionResult([
-        annif.hit.SubjectSuggestion(
+    hits1 = annif.suggestion.ListSuggestionResult([
+        annif.suggestion.SubjectSuggestion(
             uri='http://www.yso.fi/onto/yso/p10849',
             label='arkeologit',
             score=1.0)], subject_index)
     batch.evaluate(hits1, gold_set)
-    hits2 = annif.hit.ListSuggestionResult([
-        annif.hit.SubjectSuggestion(
+    hits2 = annif.suggestion.ListSuggestionResult([
+        annif.suggestion.SubjectSuggestion(
             uri='http://www.yso.fi/onto/yso/p1747',
             label='egyptologit',
             score=1.0)], subject_index)
